requirement.txt
flask>=2.3.0
flask-cors>=4.0.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.5.0
gunicorn>=21.0.0